+++ pajbot/modules/casechecker.py
@@ -200,7 +200,7 @@ class CaseCheckerModule(BaseModule):
 
         amount_uppercase = sum(1 for c in message if c.isupper())
         if self.settings["uppercase_timeouts"] is True:
-            if amount_lowercase >= self.settings["max_uppercase"]:
+            if amount_uppercase >= self.settings["max_uppercase"]:
                 self.bot.delete_or_timeout(
                     source,
                     self.settings["moderation_action"],
@@ -214,7 +214,7 @@ class CaseCheckerModule(BaseModule):
 
             if (
                 amount_uppercase >= self.settings["min_uppercase_characters"]
-                and (amount_lowercase / len(message)) * 100 >= self.settings["uppercase_percentage"]
+                and (amount_uppercase / len(message)) * 100 >= self.settings["uppercase_percentage"]
             ):
                 self.bot.delete_or_timeout(
                     source,